list all topics with segments
