+++ pkg/ensign/store/meta/topics.go
@@ -147,14 +147,15 @@ func (s *Store) ListTopics(projectID ulid.ULID) iterator.TopicIterator {
 // all topics, the entire database must be scanned looking for topic IDs as identified
 // by the topic segment.
 func (s *Store) ListAllTopics() iterator.TopicIterator {
-	// Create a range from the zero valued topic to the maximum valued topic
+	// Create a range from the zero valued object to the maximum valued object,
+	// hopefully skipping over any indexes or bigger or smaller keys.
 	slice := &util.Range{
 		Start: make([]byte, 34),
 		Limit: bytes.Repeat([]byte{0xff}, 34),
 	}
 
 	iter := s.db.NewIterator(slice, nil)
-	return &TopicIterator{iter}
+	return &TopicIterator{&SegmentIterator{iter, TopicSegment}}
 }
 
 // Create a topic in the database; if the topic already exists or if the topic is not
